Move board.js to /containers. Board is a flex row for list alignment
