+++ src/components/app.js
@@ -1,10 +1,10 @@
-import React, { Component } from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
-import Login from './login';
-import Logout from './logout';
-import Register from './register';
-import Board from './board';
-import Home from './home';
+import React, { Component } from "react";
+import { BrowserRouter as Router, Route } from "react-router-dom";
+import Login from "./login";
+import Logout from "./logout";
+import Register from "./register";
+import Board from "../containers/board";
+import Home from "./home";
 
 class App extends Component {
   render() {
